!Add DataInit(regenerated user/dept/leaveType); rearrange templates folder for Ian to deploy security.

+++ LeaveApplicationTeam5/src/main/java/sg/edu/LeaveApplication/service/UserLeaveTypesService.java
@@ -20,7 +20,7 @@ public interface UserLeaveTypesService {
 
 	public void update(User user, String leaveName, Integer leaveCost);
 	
-	public Integer findleaveAllowance(Integer userId, String leaveName);
+	public Integer findleaveAllowance(long userId, String leaveName);
 	
 	//public void updateUserLeaveAllowance(ArrayList<UserLeaveTypes> ulist);
